Create mobile ComposerInput component for DMs and comments
