+++ packages/mobile/src/components/composer-input/index.ts
@@ -1,0 +1,2 @@
+export { ComposerInput } from './ComposerInput'
+export * from './types'
